test: add testData into Cypress commands file
TIS21-3539
