+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "trainee-ui-app",
-  "version": "0.55.3",
+  "version": "0.55.4",
   "private": true,
   "dependencies": {
     "@aws-amplify/ui-react": "^2.20.0",
