+++ src/components/forms/formr-part-b/viewSections/ViewSection6.cy-test.tsx
@@ -6,11 +6,7 @@ import store from "../../../../redux/store/store";
 import ViewSection6 from "./ViewSection6";
 import history from "../../../navigation/history";
 import { FormRPartB } from "../../../../models/FormRPartB";
-import {
-  testData,
-  makeSectionEditButton,
-  formData
-} from "./ViewSectionTestHelper";
+import { makeSectionEditButton, formData } from "./ViewSectionTestHelper";
 
 type formRBSub6 = Pick<FormRPartB, "compliments">;
 
@@ -36,6 +32,6 @@ describe("View", () => {
   });
 
   it("should render correct form data", () => {
-    testData(formDataToDisplay);
+    cy.testData(formDataToDisplay);
   });
 });
