+++ src/components/forms/formr-part-b/viewSections/ViewSection7.cy-test.tsx
@@ -7,7 +7,7 @@ import store from "../../../../redux/store/store";
 import ViewSection7 from "./ViewSection7";
 import history from "../../../navigation/history";
 import { FormRPartB } from "../../../../models/FormRPartB";
-import { testData, makeSectionEditButton } from "./ViewSectionTestHelper";
+import { makeSectionEditButton } from "./ViewSectionTestHelper";
 
 let formData = submittedFormRPartBwithCovid[0];
 
@@ -35,11 +35,11 @@ describe("View", () => {
   });
 
   it("should render correct form data", () => {
-    testData(formDataToDisplay);
+    cy.testData(formDataToDisplay);
   });
 
   it("should render correct covid declaration data", () => {
-    testData(covidDeclaration);
+    cy.testData(covidDeclaration);
   });
 
   it("should show Covid training progress details", () => {
